register rtdm actions
