+++ smart_kit/resources/__init__.py
@@ -12,6 +12,8 @@ from core.basic_models.actions.counter_actions import CounterIncrementAction, Co
     CounterClearAction, CounterSetAction, CounterCopyAction
 from core.basic_models.actions.external_actions import ExternalAction
 from core.basic_models.actions.external_actions import ExternalActions
+from core.basic_models.actions.rtdm_get_pp_and_events_action import RtdmGetPpAndEventsAction
+from core.basic_models.actions.rtdm_send_response_to_pp_action import RtdmSendResponseToPpAction
 from core.basic_models.actions.string_actions import StringAction, AfinaAnswerAction, SDKAnswer, \
     SDKAnswerToUser
 from core.basic_models.answer_items.answer_items import items_factory, SdkAnswerItem, answer_items, BubbleText, \
@@ -303,6 +305,8 @@ class SmartAppResources(BaseConfig):
         actions["set_local_variable"] = SetLocalVariableAction
         actions["set_variable"] = SetVariableAction
         actions["string"] = StringAction
+        actions["rtdm_get"] = RtdmGetPpAndEventsAction
+        actions["rtdm_send"] = RtdmSendResponseToPpAction
 
     def init_requirements(self):
         requirements[None] = Requirement
